requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
djangorestframework-simplejwt==4.7.2
